+++ src/collective/weather/interfaces.py
@@ -18,6 +18,31 @@ class IWeatherUtility(Interface):
     """
 
 
+class IWeatherInfo(Interface):
+    """Communicates with a weather web service to get
+       current weather information of a given location
+    """
+
+    def getWeatherInfo(location, units, lang):
+        """Gets weather information of given location.
+
+           :param location: [required] The location from we want to get weather info.
+                Depending on the weather provider api it can represent different things:
+                a code, a city, latitude and longitude.
+           :type location: As far as the utility treats it OK it can be a string, a
+                tuple or whatever
+           :param units: Units for the returned data.
+           :type units: string, imperial/metric are the only accepted values.
+           :param lang: An ISO 639-1 lang code of 2 characters.
+           :type lang: string
+           :returns: a dictionary with these data:
+               temperature: float, expressed in the passed units.
+               summary: string with a short description of the current conditions,
+                   expressed in the passed lang.
+               icon: string with the URL of an icon representing the current conditions.
+        """
+
+
 class IWeatherSettings(form.Schema):
     """Settings for the collective.weather package.
     """